requirements.txt
fastapi
uvicorn
numpy
plotly
pathlib
concurrent-log-handler
requests
tinydb
pyperclip
cryptography
pillow
binance-futures-connector
func_timeout
pydub
aiohttp
asyncio
sourcedefender
appdirs
kaleido
opencv-python
boto3
pandas-ta
apscheduler
simpleaudio